[+BUGFIX] Extbase (Object): Disabling Dependency Injection for injectSettings method
Tx_Extbase_MVC_Controller_ActionController->resolveView() checks for the existence of 
an injectSettings() method in a view class and calls it if found, passing on the
settings.
As Tx_Extbase_Object_Container_ClassInfoFactory->getInjectMethods() though
classifies that method as a method for dependency injection which of course fails 
leading to an exception "Method is marked as setter for Dependency Injection, but 
doesnt have a type annotation".

Bugfix: Make the DI container skip methods with the name
injectSettings(). This has the benefit of making sure nobody
will use that function for injections lateron, making upgrade to
FLOW3 more easy.

Resolves: #11011