[+BUGFIX] Extbase (Utility): Make _LOCAL_LANG override work with dots
When dealing with Locallang files, it is a common approach to
write locallang keys with dots, as example tx_foo.bar.
You can overwrite these keys easily by Typoscript _LOCAL_LANG,
and with this patch this also works with keys containing dots.

Thanks to Thomas Maroschik for the initial implementation!

Resolves: #12121