[+FEATURE] Extbase: Enable Dependency Injection for Domain Models
In order to avoid the constructor to be called,
domain objects are created by using PHPs unserialize() function.
This has the side-effect, that we need to inject
dependencies "manually" for those.
This is accomplished by adding a helper function
getEmptyObject() to the Object Container.
Thanks to Pascal Jungblut for the patches!

Change-Id: I49edab54eb810abc894f3f59fe8c46cf4ddfa464
Fixes: #11311