[+FEATURE] implement ObjectManager::isRegistered()
The ObjectManager develops really fast these days. More Features
have to be back-ported. For now, especially for back-porting
the SignalSlotDispatcher, it's necessary to implement at least
isRegistered().

This is a simple implementation as we don't use scopes and stuff
in Extbase yet.

No Unit-Tests yet as there are none in FLOW3. ;-)

Change-Id: I7a31f93e9855bb1e36c5953e86c3aef7322f548f