[FEATURE] Descriptive error on invalid action name
This adds an additional check to determine if an action was specified
but does not exist in allowed actions for the current plugin.

By default, if an invalid action is specified, an InvalidActionName
Exception is thrown with the message
"There is no action <actionName> for controller <controllerName>"

If the option mvc.throwPageNotFoundExceptionIfActionCantBeResolved is
enabled (globally or just for the extension/plugin) a
t3lib_error_http_PageNotFoundException is thrown instead resulting
in a 404 error page. This is disabled by default.

The previous implementation silently used the default action if an
invalid action was specified. We consider this as a non-breaking
change nevertheless because it is very unlikely that someone was
relying on that behavior.

Thanks to Nathan L., Martin K. & Claus D. for reporting and patches!

Fixes: #5961
Change-Id: I25953da471968d7d2f34fe28d676e9de22abefc9