[FEATURE] Skip default controller and action in generated URIs
This adds a new feature that leaves out controller and/or action
arguments from generated URIs if they are equal to the default
controller/action of the target plugin.
NOTE: This feature is disabled by default because it has
unpredictable side-effects, if you override default action/controllers
via Flexform.
You can enable this with following TypoScript:
plugin.tx_<yourExtensionName>.features.skipDefaultArguments = 1

But make sure that you're not using "switchableControllerActions"
settings in your plugin FlexForms for this extension.

Change-Id: I5baf0b3f3e5470bf8eec5537af7bd5246caa1c24
Resolves: #8344