[BUGFIX] Default action is not properly resolved
When no action is specified, Extbase returns the first action
name of the default controller.
If another controller name is specified though (and no action name),
Extbase still returned the first action name of the default
controller. Instead it should obviously return the first action
name of the given controller in that case.

This is fixed by this change.

Besides invalid controller names now also trigger an exception

Change-Id: Ib4d5bbc078084206a9cdd13b0cae09926f463669
Resolves: #29875