[BUGFIX] Removing "Repository" in class name makes extensions unusable
If an extension has "Repository" in its name, it becomes useless while
determining the model name in repository class. With this patch, the
problem gets solved by removing only the last "repository" at the end
of the class name to determinate the model name from repository class.

Thanks to Kai Vogel

Change-Id: I29d52e9f6370c54d9cde09bb5e1399ce530f1659
Resolves: #30306