[BUGFIX] public function configureModule should be static
In the class Tx_Extbase_Utility_Extension is the public method configureModule
and it seems that the author of the method forget the static keyword.
The method is used as a static method everywhere but it is not declared
as static.

This patch only set the method to static.

Change-Id: I4ba19f3b0ff763dffd4109a65933623b79e85b45
Fixes: #37922
Releases: 6.0, 4.7
Reviewed-on: http://review.typo3.org/12938
Reviewed-by: Tymoteusz Motylewski
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn