[TASK] use the objectManager to create empty domain objects
The Tx_Extbase_Persistence_Mapper_DataMapper uses the objectContainer
(Tx_Extbase_Object_Container_Container) for creating an empty object
without calling the construct.

The problem is that except the objectManager no one should use the
objectContainer. So this patch adds a wrapper class to the object
manager communicates with the objectContainer.

Change-Id: I2a2f9d68c0ac0c14a47723ea901c974b7ccc6c5e
Resolves: #13816
Release: 6.0
Reviewed-on: http://review.typo3.org/12973
Reviewed-by: Felix Oertel
Tested-by: Felix Oertel