[FEATURE] Integrate basic file abstraction layer models
Warning: This feature is marked as experimental! It is a quick shot to
get a basic connection between extbase and FAL. We added the code to
see if it works out, but it is also possible that we decide to throw it
away completly and substitute it with something different later on. Be
aware of this if you use the models and type converters in your code!

Since the file abstraction layer (FAL) has been introduced during the
development of TYPO3 6.0, it is also required to have an automized and
fluent way to access those abstract files via Extbase. Since the FAL
models cannot completely and directly be used as Extbase Domain object,
simple wrapping objects are defined to access the concrete objects.

The patch adds the basic models for this. Furthermore additional type
converter classes are added for those models.

The patch contains an additional base test case that is able to create
a test database and feed it with fixture data, it is also marked as
experimental.

Change-Id: I537d32377ddd891627af2c163af252c48554d003
Resolves: #39926
Releases: 6.0
Reviewed-on: http://review.typo3.org/13885
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn