[BUGFIX] Support plugins in namespaced extensions
The function registerPlugin doesn't handle namespaced extension keys
differently than normal extension keys, leading to being unable to
read the configuration set in the configurePlugin function.
In the configurePlugin the vendor name is stripped from the extension name,
but the same is not done in registerPlugin. This leads to it being
impossible to select the plugin in the backend when inserting/editing a
plugin content element.

Change-Id: Ic21dc68e83bf65171aba7eafa43530eb7b23884b
Fixes: #41785
Releases: 6.0
Reviewed-on: http://review.typo3.org/16052
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Jan-Erik Revsbech
Tested-by: Jan-Erik Revsbech
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel