[BUGFIX] resolveValidatorObjectName support for namespaces
This fix makes sure that validators get resolved correctly for
namespaced extensions. This works for the following scenarios:
Fully qualified namespaced extension name, shorthand validator
names (namespaced & non-namespaaced) and built in validators.

This also changes the behavior for the built in validators,
so instead of resolving the deprecated class name it resolves
to the namespaced class name.

Fixes: #40955
Releases: 6.0

Change-Id: Iec6a1e613218cf1e5d032502231dfa3447d8fa96
Reviewed-on: http://review.typo3.org/15642
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel