[BUGFIX] Typo3DbBackend language handling crashes in BE
Within a past commit the language handling was completly rewritten to
support language uids other than 0. This was done for frontend only so we
have to add a backend handling as well.

The language uid is now taken from the L parameter (either POST or GET)
for backend modules. If anything was submitted a fallback to 0 is
integrated.

Change-Id: I683e4bb44fa86c2b40c6126d8360858a2319bebe
Fixes: #40796
Releases: 6.0
Reviewed-on: http://review.typo3.org/15826
Reviewed-by: Georg Ringer
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel