[BUGFIX] Reset 'vendorName' in RequestBuilder to NULL for new requests
It is currently impossible to use extensions with AND without
namespaces on the same page, since the vendorName is not reset
when a new request is built. So when the request is analyzed later on,
a vendor from a previous plugin may be found and thus the wrong
classname is built.

This patch resets the vendorName to NULL in method 'loadDefaultValues'.

Releases: 6.0
Fixes: #42782

Change-Id: Ie2d74a7140b325826fb752505e7569069f77146c
Reviewed-on: http://review.typo3.org/16481
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel