[FEATURE] Allow different overriding of object mapping per plugin
Currently the ObjectContainer has an array with a mapping of
class names to actual classes which allows the replacement
of objects via DI.

This mapping is set during the bootstrap process and uses the
typoscript settings in config.tx_extbase.objects which is too
strict. To be able to ovverride these settings for each plugin
extbase has to use 'CONFIGURATION_TYPE_FRAMEWORK' settings
instead of 'CONFIGURATION_TYPE_FULL_TYPOSCRIPT' settings.

Releases: 6.1
Resolves: #44796
Change-Id: I5d6ea135f5bec586073da3394f0ba8e0197020b7
Reviewed-on: https://review.typo3.org/17696
Reviewed-by: Mattias Nilsson
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn