[FEATURE] Allow different overriding of object mapping per plugin
Currently the ObjectContainer has an array with a mapping of
class names to actual classes which allows the replacement
of objects via DI.

This mapping is set during the bootstrap process and uses the
typoscript settings in config.tx_extbase.objects which is too
strict. To be able to ovverride these settings for each plugin
extbase has to use 'CONFIGURATION_TYPE_FRAMEWORK' settings
instead of 'CONFIGURATION_TYPE_FULL_TYPOSCRIPT' settings.

Releases: 6.1
Resolves: #43450
Change-Id: I067b19769e007578cea406964fc568dcf5727b7a
Reviewed-on: https://review.typo3.org/16855
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring