[TASK] Decouple BE module runner and Bootstrap
The Bootstrap class is currently registered as a dispatcher for backend
modules (called from mod.php). The dispatching is done in
Bootstrap::callModule(), which does not access the rest of the class
except run() (which is part of the public API).

To make the distinction more clear (and make different dispatchers
possible without logic fuckup), the module running code is moved
to a separate class ModuleRunner.

Change-Id: I7d7571d3204f9da927aa7c3d3cf0cca2b121ef15
Resolves: #45318
Releases: 6.1
Reviewed-on: https://review.typo3.org/18165
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring