[BUGFIX] TypoScript configuration not available in backend
On bootstrapping and dispatching Extbase modules in the
backend, the frontend TypoScript is being parsed. However, if
the accordant page does not have any root template assigned, the
basic Extbase settings cannot be loaded and parsed. In a result,
e.g. persistence mappings for backend users are not available in
this situation.

The solution is to explicitly trigger the processing of the
extension static files (e.g. ext_typoscript_setup.txt) even if
there's no valid root template.

Change-Id: Ibdeba3f4a0c6303989d5393fee51c20bb6741ca0
Fixes: #46511
Related: #46250
Releases: 6.0, 6.1
Reviewed-on: https://review.typo3.org/19147
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader