[!!!][BUGFIX] Introduce explicit saving of modified domain objects
In conjunction with the new property mapper the persistence layer
saves modified domain objects that failed validation and/or mapping.

This bugfix aligns the behavior of the persistence layer with Flow
and introduces explicit saving of modified domain objects.

Modified objects now need to be passed through their repository
update method in order to be scheduled for saving by the persistence
managers persistAll method.

Fixes: #47251
Releases: 6.1
Change-Id: I2130f96d925d74942af70f92e38d2d0b2ea46f79
Reviewed-on: https://review.typo3.org/19992
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring