[TASK] Check for proper extension key parts
Introduce a check for proper vendor names. If the vendor name of an
extension does not begin with a capital letter, a warning is logged
to the deprecation log.

As improper vendor names lead to errors, this check can be seen as
a first step to restrict the usage of extensions with wrong vendor
names, similar to a deprecation process.

Additionally, a refactoring of copy-pasted checks is made.

Resolves: #56944
Releases: master
Change-Id: I085f63e29c1639c0cc49fa43e225b46049d67003
Reviewed-on: http://review.typo3.org/28469
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>