[BUGFIX] Make Extbase respect sys_language_mode=strict
The Extbase query parser currently doesn't respect strict language mode
because it doesn't consider records that are only present in a localized
language version, but not in the default language.

As a consequence, records only present in a localized language version
are not fetched at all while for counting/paging the count of the default
language version is used.

This change adds a check for strict language mode and additional query
parts to solve this issue.

Releases: 6.3, 6.2
Resolves: #47183
Change-Id: I36511b2fd7c91f0bed5110c2cf12c04c8bbca9c9
Reviewed-on: https://review.typo3.org/30757
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer