[FEATURE] Introduce afterPersistObjetct signal slot
At the moment we only have an afterUpdate signal which
is emitted when an object is updated. But we don't have
a signal which is emitted after persisting an existing
aggregate root, if only sub objects or object storages
belonging to the aggregate root are updated.

Since the aggregate root should "encapsulate" the contained
objects to the outer world, we should provide such a signal,
which is fired if all objects of an aggregate root are persisted.

The same signal is emitted in the persistObject method in
the AbstractBackend class in Flow.

Change-Id: I85b2257b127f73afddbf4a01db156ed30489c42a
Resolves: #59986
Releases: 6.3
Reviewed-on: https://review.typo3.org/31238
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters