[BUGFIX] Fix Extbase language fallback in query parser
Currently when TYPO3 is configured to do language fallback
and a record is translated in language A, but not
in language B and language B is requested, then this
record is excluded.

This is the case because of a wrong subselect condition.

Extbase selects records in the requested translation
or if no translation is available in the default language.
However the check if translation is available looks
for *any* translation not only for a translation in the
requested language. Thus the record from the default
language is not selected at all if there are translations
available in any other language.

Solution is to change the subselect condition to check
for the currently requested language.

Releases: 6.1, 6.2, master
Resolves: #60613
Change-Id: I8ebd68e1f5741d3557910ae2f8c2d19474548d01
Reviewed-on: http://review.typo3.org/31859
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>