[BUGFIX] Fix things that are broken without compatiblity6
While moving things to the compatibility6 extension
it was forgotten that some aliases are required for
ext form to work properly.

Besides that, the fact that unit tests run with all
extensions enabled, it was hidden that some tests
still relied on some aliases.

These things are now all fixed and the compatibility
extension is removed from travis before executing the tests.

Resolves: #63576
Resolves: #63577
Releases: master
Change-Id: I06b50c093091dc8866a5ea8b2919565a1a5bed21
Reviewed-on: http://review.typo3.org/35070
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>