[CLEANUP] Add missing empty lines after namespace statements
At the moment in some classes you can find an empty line between the
copyright comment and the namespace statement. In some other classes
there is no such empty line.

This commit adds an empty line after the namespace statement if there
is no empty line.

This is not part of the TYPO3 CMS CGL but increases consistency in this
area.

This regular expression was used to add the empty lines:

Search:
(namespace .*\n)(/\*)

Replace:
$1\n$2

Resolves: #63831
Releases: master
Change-Id: I14c08c64ccbdb39001a5c303c629b1dfaa3a6c0f
Reviewed-on: http://review.typo3.org/35392
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>