[BUGFIX] Extbase: Cast sys_language_uid to integer in updateObject()
Inserting a new object into the database is a two step process in
extbase where the first step is creating a mostly blank record to obtain
the uid for the new record and the second step is creating all related
objects updating the initial record with the final values afterwards.
Due to this process the languageUid in the freshly created object can be
NULL, which is then used without a cast in updateObject(). If the
database enforces not inserting NULL values into columns defined as NOT
NULL this leads to SQL errors.

This is fixed by casting the languageUid to int before using it.

Resolves: #69088
Releases: master
Change-Id: I0c4519520ccc312f68a025a21d3b040c07446992
Reviewed-on: http://review.typo3.org/42653
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>