[BUGFIX] Don't convert labels for charsets in Extbase/Fluid
When Extbase/Fluid initially were introduced, it was possible
to store non-UTF8-data in the database. A charset array within
CharsetConverter checked if a special language (e.g. russian)
had its own character set. The LocalizationUtility then converted
the value from the original character set to utf-8 which is not
needed anymore, as every input data is expected to be UTF-8.

Labels coming from TypoScript are now expected to be UTF-8  as well.

Resolves: #73487
Releases: master, 7.6
Change-Id: I2292c587b7709257d8efcb58f4e5c0a4c191c944
Reviewed-on: https://review.typo3.org/46711
Reviewed-by: Sebastian Fischer <typo3@evoweb.de>
Tested-by: Sebastian Fischer <typo3@evoweb.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>