[FEATURE] Enforce cHash argument for Extbase actions
TypoScriptFrontendController::reqCHash() is now called for Extbase
frontend plugin actions just like they were usually called for the
legacy AbstractPlugin.

This throws a 404, if plugin arguments are present, but cHash is not,
which would also happen, if the plugin arguments were added
to "cHashRequiredParameters" configuration.

This provides a more reliable page caching behavior
by default and with zero configuration for extension authors.

With the feature switch "requireCHashArgumentForActionArguments" this
behavior can be disabled, which could be useful, if all actions in a plugin
are uncached or one wants to manually control the cHash behavior.

Resolves: #78002
Releases: master
Change-Id: Ic6f71c3e5c8a94a0d422372a08c944aef5663f06
Reviewed-on: https://review.typo3.org/49976
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>