[BUGFIX] Remove relations from MM table
If extbase deletes an object it doesn't delete its relations in an
MM table. This patch adds the check for a possible MM relation table and
ensures all relations get deleted as well. If the parent table supports
a deleted flag, no relations will be deleted to be able to restore the
parent record again.

Resolves: #78128
Releases: master, 7.6
Change-Id: Ibe2497c05838e261f9c67a80216385d6e60607cb
Reviewed-on: https://review.typo3.org/50103
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Robert Jelinek <robert.jelinek@riir.at>
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>