[!!!][FEATURE] Introduce single _cli_ user for all CLI calls
The command line users prefixing with _CLI_scheduler etc. is flawed,
however a single user is needed for doing e.g. DataHandler calls.

This patch adds a new CommandLineUserAuthentication class, which extends
the regular $BE_USER object to only allow a user called "_cli_" and which
is created automatically if it does not exist, so the _cli_ user always
exists.

Additionally, the _cli_ user is now an admin user so access settings
on command line are obsolete and crucial stuff like the DataHandler can
be used without restrictions.

* Adapt Extbase Bootstrap
* Make sure to have an API to "opt-in" for the need for a user
* Adapt documentation (e.g. in scheduler) because a new user is not needed.

Resolves: #79240
Releases: master
Change-Id: I5a6277fa2252ce6c61f57bbfa26e8be9dc47bbab
Reviewed-on: https://review.typo3.org/51089
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>