[TASK] Disable ensureCorrectPostOrderingByCategoryTitle on postgres
A recently added test in extbase query building reveals a
non-trivial-to-fix issue of the created query with postgres.
This is the last issue we have with functionals on postgres, to
have them running as pre-merge tests we for now disable the offending
tests on that database platform, and handle the real fix in a
separate issue.

Change-Id: I830af04ecdb41fa968421521e9860c388f74d75b
Resolves: #79987
Related: #79988
Related: #79931
Related: #79932
Related: #79330
Releases: master
Reviewed-on: https://review.typo3.org/51818
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>