[BUGFIX] mssql: ExpressionBuilder inSet() support
mssql does not support FIND_IN_SET(). The patch adds a
solution based on LIKE.
Since the query fiddling in this area is a bit tricky,
this area is now supported by a bunch of functional tests.
A postgres bug those new functional tests reveal is fixed
along the way.

Change-Id: I5e94ad8df7a37a680b457eff1b5b16a0c14dba39
Resolves: #81488
Releases: master, 8.7
Reviewed-on: https://review.typo3.org/53147
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>