[TASK] Remove unused Extbase exception classes
There are a lot PHP exception classes which are not in use inside Extbase.

Some come from the fact that Fluid got more mature and
standalone, some from refactorings within ObjectManager,
PropertyManager, RequestHash and Persistence.

One highlight is the InvalidMarkerException which dates back
to the backport of Flow before Fluid was invented and Flow
was using the same marker-based approach as TYPO3 v4 had -
and was never used in TYPO3 v4 at all.

Class aliases are in place for all of them, so no breaking change.

Resolves: #81600
Releases: master
Change-Id: Ic1c3fc309954750f8c2e664ab94fa0e6f9336bea
Reviewed-on: https://review.typo3.org/53224
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>