[BUGFIX] Do not detect single select fields as relation
With https://review.typo3.org/50879/ all TCA select fields
are detected as relation to many, even when they are configured
as simple static select.

This leads to the row value to be pre-set as string (to hold comma
separated values), which leads to an exception in MySQL Strict Mode
when the select value field is defined as integer.

Therefore only detect renderType === selectMultipleSideBySide
as relational field

Resolves: #81043
Releases: master, 8.7
Change-Id: I54effaace3eb19034d86a9a74bbd5cc207e867f1
Reviewed-on: https://review.typo3.org/53565
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>