[TASK] Functional tests for translated tt_content rendering in Extbase
Introduce and Extbase version of the LocalizedContentRenderingTest.
It tests how language related TS settings influences content rendering.

Also changes exception code in ContentController, to make bamboo happy.

This test suite is documenting current behaviour, which is in fact
inconsistent with TypoScript tt_content rendering.

The next step is to provide a fix for extbase with a feature switch,
so this test suite will become a version with feature swich being disabled.

Resolves: #81823
Releases: master, 8.7, 7.6
Change-Id: Ia616e5089b0209cfb895216726f7192f39c88a45
Reviewed-on: https://review.typo3.org/53426
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>