[!!!][TASK] Migrate modules to regular backend routing
This patch removes the separate request handler for backend modules,
which was accessed via "&M=moduleName" GET parameter. This is now
migrated into the RouteDispatcher which can dispatch modules as well.

Now, modules are called via the "&route" parameter like all other routes.

Additionally, the requested URLs for modules were requested with the additional
"moduleToken" which is now called "token".

This way, special treatment for modules when dispatching is removed,
however the security checks are still in place so this is kept as is.

All places where URLs are generated can now still be accessed via
`BackendUtility::getModuleUrl()` which can deal with routes, module names
and routePaths (from the URL) to keep backwards-compatibility.

Next Steps:
- Migration wizard for bookmarks + Streamline bookmarks code (see todos)
- Check what needs to be added in ExtensionManagementUtility
- Introduce slugs in routes for BE, e.g. /file-edit/{fileId}/ and /module/page/view/{id}
- Document reserved GET parameters "id", "route" and "token"
- Cleanup usage of determineScriptId and getModuleUrl to use new API

Resolves: #82406
Releases: master
Change-Id: If11c3d5289e14bc9ea766468b8e94cce95c23c71
Reviewed-on: https://review.typo3.org/53881
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Matthias Vogel <typo3@kanti.de>
Tested-by: Matthias Vogel <typo3@kanti.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>