[BUGFIX] Remove duplicate PageTS cache in Typo3DbBackend
The Typo3DbBackend fetches PageTS via BackendUtility::getPagesTSconfig
and then stores a copy of it (for each page separately) in a static
property. This is a bug and leads to performance decreases because
the PageTS is already cached via BackendUtility::getPagesTSconfig
and it's done even better than in Typo3DbBackend because the backend
utility store a single cache entry for different pages with the same
PageTS. The Typo3DbBackend does create a copy of the PageTS for each
page.

Resolves: #82496
Releases: master, 8.7
Change-Id: I61bd6fa8eb3ba29baea355ae01a8d71f7ae7e963
Reviewed-on: https://review.typo3.org/54185
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>