[FEATURE] Replace @inject with @TYPO3\CMS\Extbase\Annotation\Inject
This patch introduces the usage of doctrine annotations.
Therefore the AnnotationRegistry is configured during the
bootstrap and the annotation "TYPO3\CMS\Extbase\Annotation\Inject"
is evaluated when building a ClassSchema for a class.

On top of that this patch puts the static variable
$ignoredTags into the DocCommentParser class, which
saved quite some bytes of RAM when reconstituting
the ClassSchema objects from the cache.

Resolves: #82869
Releases: master
Change-Id: I99d17706395ebc0c308c7f611c59f207fe0b6233
Reviewed-on: https://review.typo3.org/54590
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>