[BUGFIX] fix missing l18n fields in blog tag table
The field l18n_parent was missing in table
tx_blogexample_domain_model_tag, but configured in TCA.
This causes an exception with adding a new tag in BE.
Also the field l18n_diffsource was missing in TCA and database.

* Add the fields l18n_parent and l18_diffsource to
  the databse table tx_blogexample_domain_model_tag
* Add TCA configuration for field l18n_diffsource to
  tx_blogexample_domain_model_tag TCA

Also add missing icon and label for
tx_blogexample_domain_model_dateexample.

Remove "allow tables on standard pages". This is not
necessary.

Fixes: #83113
Releases: master, 8.7
Change-Id: I10f809cc4b00df07c6e5a064bb89018578f3abf3
Reviewed-on: https://review.typo3.org/54786
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Hölzle <typo3@markus-hoelzle.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Łukasz Uznański <l.uznanski@macopedia.pl>
Tested-by: Łukasz Uznański <l.uznanski@macopedia.pl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>