[BUGFIX] get count consistent with execute for self written statements
Fix wrong result for self written sql statements executing count
on the query result.

To get this consistent, check for statement also in
getObjectCountByQuery and run getObjectDataByQuery
for self written statements. After that count the
rows.

Wrapping the statement in a subselect will
may be break doLanguageAndWorkspaceOverlay
in getObjectDataByQuery.

This is just to be consistent. Normally it does not
make sense to call the count method on a self written
statement.

Releases: master, 8.7
Resolves: #82910
Change-Id: I8bddb1b20d062d9f46939c16885f1eb73df1e5f4
Reviewed-on: https://review.typo3.org/54844
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>