[BUGFIX] Allow dev branch install of subtree split packages
To make it possible to require subtree split packages
as a dev 8.7.*@dev versions, we need to adapt the
dependencies of the packages.

For a release these dependencies will be set to
exact version numbers before tagging.

Branch aliases in the packages can be removed,
as the TYPO3_8-7 branch does not exist in the
splitted repos, but only a 8.7 branch.

Resolves: #83671
Releases: 8.7
Change-Id: I2784123f596409b720a090276c4314fd1a1c4a63
Reviewed-on: https://review.typo3.org/55444
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>